requirements.txt
cmake
djangorestframework
python-dateutil
Pyjwt
Django==2.2.4
django-cors-headers
nose
Sphinx==1.5.5
sphinx-rtd-theme==0.2.4
openpyxl==2.5.6
django-crontab==0.7.1
Pillow==5.2.0
numpy==1.17.2
Click==7.0
opencv-python==4.4.0.44